games/beast/pyproject.toml
[project]
name = "Beast"
version = "0.1.0"
description = "Catch falling stars before they hit the ground!"
requires-python = ">=3.10,<4.0"
dependencies = [
    "retro-games==2.2.0",
]

[project.scripts]
play = "beast:main"

[tool.retro]
name = "Beast"
author = "Chris"
description = "Push blocks to squish the beasties. Don't let them catch you!"
instructions = "Use the arrow keys to push blocks. Squish beasts between blocks."
result_file = "result.json"
log_file = "game.log"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
